Feat: 상품 상세 페이지 로그인 여부에 따른 버튼 설정 #289

+++ agijagi/app/src/main/java/likelion/project/agijagi/product/ProductDetailFragment.kt
@@ -1,6 +1,7 @@
 package likelion.project.agijagi.product
 
 import android.os.Bundle
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -13,6 +14,8 @@ import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
 import com.google.firebase.storage.ktx.storage
 import io.supercharge.shimmerlayout.ShimmerLayout
+import likelion.project.agijagi.MainActivity.Companion.displayDialogUserNotLogin
+import likelion.project.agijagi.MainActivity.Companion.userModel
 import likelion.project.agijagi.R
 import likelion.project.agijagi.databinding.FragmentProductDetailBinding
 import java.text.DecimalFormat
@@ -149,9 +152,13 @@ class ProductDetailFragment : Fragment() {
                 findNavController().popBackStack()
             }
             setOnMenuItemClickListener {
-                when (it.itemId) {
-                    R.id.menu_product_detail_shopping -> {
-                        findNavController().navigate(R.id.action_productDetailFragment_to_shoppingListFragment)
+                if (userModel.uid == "") {
+                    displayDialogUserNotLogin(requireContext())
+                } else {
+                    when (it.itemId) {
+                        R.id.menu_product_detail_shopping -> {
+                            findNavController().navigate(R.id.action_productDetailFragment_to_shoppingListFragment)
+                        }
                     }
                 }
                 false
@@ -162,44 +169,56 @@ class ProductDetailFragment : Fragment() {
     private fun setupFavoriteButton(productId: String) {
         val buyerId = "Ws3TxAyKAg6Xe5GVNwV4"
         binding.imageButtonProductDetailFavorite.run {
-            db.collection("buyer")
-                .document(buyerId)
-                .collection("wish")
-                .get()
-                .addOnSuccessListener {
-                    for (document in it) {
-                        if (document.id == productId) {
-                            this.isSelected = true
+            if (userModel.uid != "") {
+                db.collection("buyer")
+                    .document(buyerId)
+                    .collection("wish")
+                    .get()
+                    .addOnSuccessListener {
+                        for (document in it) {
+                            if (document.id == productId) {
+                                this.isSelected = true
+                            }
                         }
+                    }.addOnFailureListener {
+                        Log.d("hye", it.toString())
+                        Log.d("hye", userModel.roleId)
+                    }
+            }
+
+            setOnClickListener {
+                if (userModel.uid == "") {
+                    displayDialogUserNotLogin(requireContext())
+                } else {
+                    it.isSelected = it.isSelected != true
+                    if (it.isSelected) {
+                        val prodId = hashMapOf("prodId" to productId)
+                        db.collection("buyer")
+                            .document(buyerId)
+                            .collection("wish")
+                            .document(productId)
+                            .set(prodId)
+                    } else {
+                        db.collection("buyer")
+                            .document(userModel.roleId)
+                            .collection("wish")
+                            .document(productId)
+                            .delete()
                     }
                 }
-            setOnClickListener {
-                it.isSelected = it.isSelected != true
-                if (it.isSelected) {
-                    val prodId = hashMapOf("prodId" to productId)
-                    db.collection("buyer")
-                        .document(buyerId)
-                        .collection("wish")
-                        .document(productId)
-                        .set(prodId)
-
-                } else {
-                    db.collection("buyer")
-                        .document(buyerId)
-                        .collection("wish")
-                        .document(productId)
-                        .delete()
-                }
-
             }
         }
     }
 
     private fun setupPurchaseButton(productId: String) {
         binding.buttonProductDetailPurchase.setOnClickListener {
-            val bundle = bundleOf("prodId" to productId)
-            it.findNavController()
-                .navigate(R.id.action_productDetailFragment_to_readyMadeOptionFragment, bundle)
+            if (userModel.uid == "") {
+                displayDialogUserNotLogin(requireContext())
+            } else {
+                val bundle = bundleOf("prodId" to productId)
+                it.findNavController()
+                    .navigate(R.id.action_productDetailFragment_to_readyMadeOptionFragment, bundle)
+            }
         }
     }
